% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_target_order}
\alias{add_target_order}
\title{Add the order column to the control target}
\usage{
add_target_order(control)
}
\arguments{
\item{control}{A fwdControl object}
}
\value{
A fwdControl object with an order column.
}
\description{
Add the order column to the control target data.frame so that targets are processed in the correct order.
}
\details{
It is important that the targets in the control object are processed in the correct order.
Targets can happen simultaneously. For example, if there are multiple FLFishery objects in
operating model each will need to have a target to solve for at the same time as the others.
The targets are processed in a time ordered sequence (year / season).
However, within the same year and season it is necessary for the min and max targets to be processed
separatley and after the other targets.
}
