% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcb_draw.R
\docType{class}
\name{basicBlock}
\alias{basicBlock}
\alias{basicBlock-class}
\title{A base class for drawing FCB bits}
\description{
Look up and see the flying saucers cruising in the sky
I saw one myself it ain't no lie
Look down and see the road you're on as if you are on a marathon
That's the spirit, victory or die
}
\section{Slots}{

    \describe{
    \item{height}{height of box}
    \item{width}{width of box}
    \item{x_centre}{x coordinate of centre of box}
    \item{y_centre}{y coordinate of centre of box}
    \item{name}{text to go in the box}
    \item{name_cex}{size of text to go in the box}
}
}

\author{
Finlay Scott - EC JRC.
}
\keyword{classes}
