% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcb_draw.R
\docType{class}
\name{biolBlock}
\alias{biolBlock}
\alias{biolBlock-class}
\title{A class for drawing a biological stock}
\description{
A class for drawing a biological stock
}
\section{Slots}{

    \describe{
    \item{height}{height of box}
    \item{width}{width of box}
    \item{x_centre}{x coordinate of centre of box}
    \item{y_centre}{y coordinate of centre of box}
    \item{name}{text to go in the box}
    \item{name_cex}{size of text to go in the box}
    \item{neck_length}{neck length}
    \item{circle_cex}{size of the circle bit}
}
}

\author{
Finlay Scott - EC JRC.
}
\keyword{classes}
