% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_F}
\alias{calc_F}
\title{Calculate fishing mortality}
\usage{
calc_F(catch, biol, effort)
}
\arguments{
\item{catch}{The FLCatch}

\item{biol}{The FLBiol}

\item{effort}{The fishing effort}
}
\description{
Calculate F in the same way as the internal C++ does
}
