% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{coerce}
\alias{coerce}
\title{Methods for coercing objects between classes}
\arguments{
\item{from}{Object to be coerced into one of another class.}

\item{to}{Name of the output class, \emph{character}.}
}
\value{
An object of the requested class.
}
\description{
A call to \emph{as(from, 'to')} will coerce the object \emph{from}, of a certain class,
to one of class \emph{to}, as specified in the method.
}
\details{
An object of class \emph{FLQuants} can be coerced into a \emph{fwdControl}, through a
call to \emph{as.data.frame}. The name of the element, or elements, in the object
specifies the 'quant' in \emph{fwdControl}. The 'quant' in the \emph{FLQuant} object, the
name of the first dimension, is ignored unles is one of 'min', 'value' or 'max'.
See the examples below on how to pass one or more \emph{FLQuant} objects to \emph{fwd}.
}
\examples{
# Single *catch* target
as(FLQuants(catch=FLQuant(4500, dimnames=list(year=2000))), "fwdControl")
# Single single *f* range
as(FLQuants(f=FLQuant(c(0.1, 0.9),
  dimnames=list(quant=c("min", "max"), year=2000))), 'fwdControl')
# Single *f* target, *value* specified
as(FLQuants(f=FLQuant(0.5, dimnames=list(quant=c("value"), year=2000))),
  'fwdControl')
# *catch* and *ssb* targets
as(FLQuants(catch=FLQuant(4500, dimnames=list(year=2000)),
   ssb_end=FLQuant(12000, dimnames=list(year=2000))), "fwdControl")
# *f* target and *catch* limits
as(FLQuants(f=FLQuant(0.5, dimnames=list(year=2000)),
  catch=FLQuant(c(100, 4000), dimnames=list(quant=c("min", "max"), year=2000))),
  'fwdControl')
# *f* target and *catch* minimum
as(FLQuants(f=FLQuant(0.5, dimnames=list(year=2000)),
  catch=FLQuant(c(100), dimnames=list(quant=c("min"), year=2000))), 'fwdControl')
}
\seealso{
\link{coerce}
}
\author{
Iago Mosqueira. EC JRC.
}
\keyword{methods}
