% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcb_draw.R
\docType{methods}
\name{draw}
\alias{draw}
\alias{draw-method}
\alias{draw,basicBlock-method}
\alias{draw-basicBlock}
\alias{draw,biolBlock-method}
\alias{draw-biolBlock}
\alias{draw,fisheryBlock-method}
\alias{draw-fisheryBlock}
\alias{draw,catchBlock-method}
\alias{draw-catchBlock}
\alias{draw,FCBDrawing-method}
\alias{draw-FCBDrawing}
\alias{draw,matrix-method}
\alias{draw-matrix}
\title{Generic method for drawing components the FCB matrix}
\usage{
draw(object)

\S4method{draw}{basicBlock}(object)

\S4method{draw}{biolBlock}(object)

\S4method{draw}{fisheryBlock}(object)

\S4method{draw}{catchBlock}(object)

\S4method{draw}{FCBDrawing}(object)

\S4method{draw}{matrix}(object)
}
\arguments{
\item{object}{The object - link{basicBlock}, link{biolBlock}, link{fisheryBlock}, link{catchBlock}, link{matrix}, link{FCBDrawing}}
}
\value{
Nothing. Just draws something
}
\description{
Used to draw individual components or the whole FCB matrix

Method for drawing the basic block of the FCB matrix

Method for drawing the biological block of the FCB matrix

Method for drawing the fishery block of the FCB matrix

Method for drawing the catch block of the FCB matrix

Method for drawing the \link{FCBDrawing} object

Method for drawing the FCB matric
}
\examples{
FCB <- matrix(c(1,1,1,1,2,2,2,1,2,2,2,3,2,2,4), nrow=5, byrow=TRUE)
draw(FCB)
}
