% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{fcb}
\alias{fcb}
\title{Generate the FCB matrix}
\usage{
fcb(biols, fisheries)
}
\arguments{
\item{biols}{The FLBiols.}

\item{fisheries}{The FLFisheries.}
}
\description{
Tries to generate FCB matrix based on names.
Internal function. Ignore.
}
