% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{fcb2int}
\alias{fcb2int}
\title{fcb2int function}
\usage{
fcb2int(fcb, biols, fisheries)
}
\arguments{
\item{fcb}{The FCB matrix}

\item{biols}{The biols}
}
\description{
Internal function not for public consumption
}
