% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcb_draw.R
\docType{class}
\name{FCBDrawing}
\alias{FCBDrawing}
\alias{FCBDrawing-class}
\alias{FCBDrawing}
\alias{FCBDrawing-method}
\alias{FCBDrawing,matrix-method}
\alias{FCBDrawing-matrix}
\title{A class for drawing the FCB matrix}
\usage{
FCBDrawing(FCB, ...)

\S4method{FCBDrawing}{matrix}(FCB, fisheryNames = NULL, catchNames = NULL,
  biolNames = NULL)
}
\arguments{
\item{FCB}{The (FCB \link{matrix})}

\item{...}{Other arguments}

\item{fisheryNames}{A vector of names of the FLFishery blocks}

\item{catchNames}{A vector of names of the FLCatch blocks}

\item{biolNames}{A vector of names of the FLBiol blocks}
}
\value{
An \link{FCBDrawing} object
}
\description{
A class for drawing the FCB matrix

Generic constructor for making an \link{FCBDrawing} object

Has to figure out the height, width and positions of the FCB components
}
\section{Slots}{

    \describe{
    \item{biolBlocks}{A list of \link{biolBlock}s}
    \item{fisheryBlocks}{A list of \link{fisheryBlock}s}
    \item{catchBlocks}{A list of \link{catchBlock}s}
    \item{CBConnectors}{A list of coordinates connecting catches and biols}
    \item{FCB}{The FCB matrix}
}
}

\author{
Finlay Scott - EC JRC.
}
\keyword{classes}
