% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utilities.R
\name{fillchar}
\alias{fillchar}
\alias{fillchar,FLFisheries-method}
\alias{fillchar,FLBiols-method}
\title{Fill up character slots}
\usage{
fillchar(object)

\S4method{fillchar}{FLFisheries}(object)

\S4method{fillchar}{FLBiols}(object)
}
\arguments{
\item{object}{The object (FLFisheries or FLBiols)}
}
\description{
Fill the character slots (name etc) with something.
}
