% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcb_draw.R
\docType{class}
\name{fisheryBlock}
\alias{fisheryBlock}
\alias{fisheryBlock-class}
\title{A class for drawing a fishery}
\description{
A class for drawing a fishery
}
\section{Slots}{

    \describe{
    \item{height}{height of box}
    \item{width}{width of box}
    \item{x_centre}{x coordinate of centre of box}
    \item{y_centre}{y coordinate of centre of box}
    \item{name}{text to go in the box}
    \item{name_cex}{size of text to go in the box}
    \item{neck_length}{neck length}
    \item{no_tails}{number of catches}
    \item{tail_gap}{space between catches}
    \item{tail_length}{tail length}
}
}

\author{
Finlay Scott - EC JRC.
}
\keyword{classes}
