% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{flasher_annual_stock_prep}
\alias{flasher_annual_stock_prep}
\title{Creates an FLFisheries and biol bits for use in FLasher from an FLStock}
\usage{
flasher_annual_stock_prep(stk, srr_model, srr_params, srr_residuals,
  srr_residuals_mult, niters = 1)
}
\arguments{
\item{stk}{An annual FLStock.}

\item{srr_model}{Name of the stock-recruitmen model.}

\item{srr_params}{Parameters for the SRR as an FLPar (i.e. directly from an FLSR).}

\item{srr_residuals}{An FLQuant of residuals for the SRR.}

\item{srr_residuals_mult}{Boolean to determine of the residuals are multiplicative or additive.}
}
\value{
A list of objects for use with FLasher.
}
\description{
Creates the bits needed to call FLasher from an annual FLStock.
Used for testing purposes.
Careful with hperiod. Not able to calculate from m.spwn and harvest.spwn slots
so F is assumed to happen continuously through year (like m).
}

