% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_FLFishery_empty_constructor}
\alias{test_FLFishery_empty_constructor}
\alias{test_simple_FLFishery_sexp_constructor}
\alias{test_FLFishery_sexp_constructor}
\alias{test_FLFishery_as_wrap}
\alias{test_FLFishery_copy_constructor}
\alias{test_FLFishery_assignment_operator}
\alias{test_FLFishery_const_catches_get_accessors}
\alias{test_FLFishery_catches_get_accessors}
\alias{test_FLFishery_const_economics_get_accessors}
\alias{test_FLFishery_const_get_hperiod}
\alias{test_FLFishery_get_hperiod}
\alias{test_FLFishery_set_hperiod}
\alias{test_FLFishery_get_effort_subset}
\alias{test_FLFishery_get_effort}
\alias{test_FLFishery_economics_get_accessors}
\alias{test_FLFishery_set_accessors}
\alias{test_FLFishery_copy_constructor2}
\alias{test_FLFishery_assignment_operator2}
\alias{test_FLFisheryAD_empty_constructor}
\alias{test_simple_FLFisheryAD_sexp_constructor}
\alias{test_FLFisheryAD_sexp_constructor}
\alias{test_FLFisheryAD_as_wrap}
\alias{test_FLFisheryAD_copy_constructor}
\alias{test_FLFisheryAD_assignment_operator}
\alias{test_FLFisheryAD_const_catches_get_accessors}
\alias{test_FLFisheryAD_catches_get_accessors}
\alias{test_FLFisheryAD_const_economics_get_accessors}
\alias{test_FLFisheryAD_economics_get_accessors}
\alias{test_FLFisheryAD_set_accessors}
\alias{test_FLFisheryAD_copy_constructor2}
\alias{test_FLFisheryAD_assignment_operator2}
\alias{test_FLFisheryAD_const_iterator}
\alias{test_FLFisheryAD_iterator}
\alias{test_FLFisheries_sexp_constructor}
\alias{test_FLFisheries_as_wrap}
\alias{test_FLFisheries_get_nfisheries}
\alias{test_FLFisheries_copy_constructor}
\alias{test_FLFisheries_copy_constructor2}
\alias{test_FLFisheries_assignment_operator}
\alias{test_FLFisheries_assignment_operator2}
\alias{test_FLFisheries_const_get_single}
\alias{test_FLFisheries_get_single}
\alias{test_FLFisheries_const_get_double}
\alias{test_FLFisheries_get_double}
\alias{test_FLFisheries_set_single}
\alias{test_FLFisheries_set_double}
\alias{test_FLFishery_revenue}
\alias{test_FLFishery_revenue_subset}
\alias{test_FLFisheriesAD_sexp_constructor}
\alias{test_FLFisheriesAD_as_wrap}
\alias{test_FLFisheriesAD_get_nfisheries}
\alias{test_FLFisheriesAD_copy_constructor}
\alias{test_FLFisheriesAD_copy_constructor2}
\alias{test_FLFisheriesAD_assignment_operator}
\alias{test_FLFisheriesAD_assignment_operator2}
\alias{test_FLFisheriesAD_const_get_single}
\alias{test_FLFisheriesAD_get_single}
\alias{test_FLFisheriesAD_const_get_double}
\alias{test_FLFisheriesAD_get_double}
\alias{test_FLFisheriesAD_set_single}
\alias{test_FLFisheriesAD_set_double}
\alias{test_FLFisheriesAD_const_iterator}
\alias{test_FLFisheriesAD_iterator}
\title{Shit tonnes of them!}
\usage{
test_FLFishery_empty_constructor()

test_simple_FLFishery_sexp_constructor(flf_sexp)

test_FLFishery_sexp_constructor(flf_sexp)

test_FLFishery_as_wrap(flf)

test_FLFishery_copy_constructor(flf1)

test_FLFishery_assignment_operator(flf1)

test_FLFishery_const_catches_get_accessors(flf, catches, quant, year, unit,
  season, area, iter)

test_FLFishery_catches_get_accessors(flf, catches, quant, year, unit, season,
  area, iter)

test_FLFishery_const_economics_get_accessors(flf, year, unit, season, area,
  iter)

test_FLFishery_const_get_hperiod(flf, quant, year, unit, season, area, iter)

test_FLFishery_get_hperiod(flf, quant, year, unit, season, area, iter)

test_FLFishery_set_hperiod(flf, quant, year, unit, season, area, iter, value)

test_FLFishery_get_effort_subset(flf, indices_min, indices_max)

test_FLFishery_get_effort(flf)

test_FLFishery_economics_get_accessors(flf, year, unit, season, area, iter)

test_FLFishery_set_accessors(flf, catches, quant, year, unit, season, area,
  iter, value)

test_FLFishery_copy_constructor2(flf1, element, quant, year, unit, season, area,
  iter, value)

test_FLFishery_assignment_operator2(flf1, element, quant, year, unit, season,
  area, iter, value)

test_FLFisheryAD_empty_constructor()

test_simple_FLFisheryAD_sexp_constructor(flf_sexp)

test_FLFisheryAD_sexp_constructor(flf_sexp)

test_FLFisheryAD_as_wrap(flf)

test_FLFisheryAD_copy_constructor(flf1)

test_FLFisheryAD_assignment_operator(flf1)

test_FLFisheryAD_const_catches_get_accessors(flf, catches, quant, year, unit,
  season, area, iter)

test_FLFisheryAD_catches_get_accessors(flf, catches, quant, year, unit, season,
  area, iter)

test_FLFisheryAD_const_economics_get_accessors(flf, year, unit, season, area,
  iter)

test_FLFisheryAD_economics_get_accessors(flf, year, unit, season, area, iter)

test_FLFisheryAD_set_accessors(flf, catches, quant, year, unit, season, area,
  iter, value)

test_FLFisheryAD_copy_constructor2(flf1, element, quant, year, unit, season,
  area, iter, value)

test_FLFisheryAD_assignment_operator2(flf1, element, quant, year, unit, season,
  area, iter, value)

test_FLFisheryAD_const_iterator(fishery)

test_FLFisheryAD_iterator(fishery, quant, year, unit, season, area, iter, value)

test_FLFisheries_sexp_constructor(flfs_sexp1)

test_FLFisheries_as_wrap(flfs)

test_FLFisheries_get_nfisheries(flfs)

test_FLFisheries_copy_constructor(flfs)

test_FLFisheries_copy_constructor2(flfs1, fishery, catches, indices, value)

test_FLFisheries_assignment_operator(flfs)

test_FLFisheries_assignment_operator2(flfs1, fishery, catches, indices, value)

test_FLFisheries_const_get_single(flfs, fishery, catches, indices)

test_FLFisheries_get_single(flfs, fishery, catches, indices)

test_FLFisheries_const_get_double(flfs, fishery, catches, indices)

test_FLFisheries_get_double(flfs, fishery, catches, indices)

test_FLFisheries_set_single(flfs, fishery, catches, indices, value)

test_FLFisheries_set_double(flfs, fishery, catches, indices, value)

test_FLFishery_revenue(flf)

test_FLFishery_revenue_subset(flf, indices_min, indices_max)

test_FLFisheriesAD_sexp_constructor(flfs_sexp1)

test_FLFisheriesAD_as_wrap(flfs)

test_FLFisheriesAD_get_nfisheries(flfs)

test_FLFisheriesAD_copy_constructor(flfs)

test_FLFisheriesAD_copy_constructor2(flfs1, fishery, catches, indices, value)

test_FLFisheriesAD_assignment_operator(flfs)

test_FLFisheriesAD_assignment_operator2(flfs1, fishery, catches, indices, value)

test_FLFisheriesAD_const_get_single(flfs, fishery, catches, indices)

test_FLFisheriesAD_get_single(flfs, fishery, catches, indices)

test_FLFisheriesAD_const_get_double(flfs, fishery, catches, indices)

test_FLFisheriesAD_get_double(flfs, fishery, catches, indices)

test_FLFisheriesAD_set_single(flfs, fishery, catches, indices, value)

test_FLFisheriesAD_set_double(flfs, fishery, catches, indices, value)

test_FLFisheriesAD_const_iterator(fisheries)

test_FLFisheriesAD_iterator(fisheries, value)
}
\arguments{
\item{flf_sexp}{something}

\item{flf}{something}

\item{flf1}{something}

\item{catches}{something}

\item{quant}{something}

\item{year}{something}

\item{unit}{something}

\item{season}{something}

\item{area}{something}

\item{iter}{something}

\item{value}{something}

\item{indices_min}{something}

\item{indices_max}{something}

\item{element}{something}

\item{fishery}{something}

\item{flfs_sexp1}{something}

\item{flfs}{something}

\item{flfs1}{something}

\item{indices}{something}

\item{fisheries}{something}
}
\description{
Shit tonnes of them!
}
