% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_FLQuantAD_as_wrap}
\alias{test_FLQuantAD_as_wrap}
\alias{test_FLQuantAD_basic_constructor}
\alias{test_FLQuantAD_sexp_constructor}
\alias{test_FLQuantAD_dim_constructor}
\alias{test_FLQuantAD_copy_constructor}
\alias{test_FLQuantAD_copy_constructor2}
\alias{test_FLQuantAD_assignment_operator}
\alias{test_FLQuantAD_assignment_operator2}
\alias{test_FLQuantAD_from_FLQuant_constructor}
\alias{test_FLQuantAD_get_units}
\alias{test_FLQuantAD_get_dim}
\alias{test_FLQuantAD_get_dimnames}
\alias{test_FLQuantAD_get_size}
\alias{test_FLQuantAD_get_nquant}
\alias{test_FLQuantAD_get_nyear}
\alias{test_FLQuantAD_get_nunit}
\alias{test_FLQuantAD_get_nseason}
\alias{test_FLQuantAD_get_narea}
\alias{test_FLQuantAD_get_niter}
\alias{test_FLQuantAD_get_data_element}
\alias{test_FLQuantAD_get_const_single_index_accessor}
\alias{test_FLQuantAD_get_single_index_accessor}
\alias{test_FLQuantAD_set_single_index_accessor}
\alias{test_FLQuantAD_const_get_accessor}
\alias{test_FLQuantAD_get_accessor}
\alias{test_FLQuantAD_set_accessor}
\alias{test_FLQuantAD_get_const_indices_accessor}
\alias{test_FLQuantAD_get_indices_accessor}
\alias{test_FLQuantAD_set_indices_accessor}
\alias{test_FLQuantAD_set_dimnames}
\alias{test_FLQuantAD_FLQuantAD_match_dims}
\alias{test_FLQuantAD_FLQuant_match_dims}
\alias{test_FLQuant_FLQuantAD_match_dims}
\alias{test_FLQuantAD_fill}
\alias{test_FLQuantAD_fill_double}
\alias{testFLCppAD}
\alias{FLPar_units}
\alias{test_FLQuant_as_wrap}
\alias{test_FLQuant_basic_constructor}
\alias{test_FLQuant_sexp_constructor}
\alias{test_FLQuant_dim_constructor}
\alias{test_FLQuant_dim_value_constructor}
\alias{test_FLQuant_intvec_dim_constructor}
\alias{test_FLQuant_intvec_dim_value_constructor}
\alias{test_FLQuant_copy_constructor}
\alias{test_FLQuant_copy_constructor2}
\alias{test_FLQuant_assignment_operator}
\alias{test_FLQuant_assignment_operator2}
\alias{test_FLQuant_from_FLQuantAD_constructor}
\alias{test_FLQuant_get_data}
\alias{test_FLQuant_get_units}
\alias{test_FLQuant_get_dim}
\alias{test_FLQuant_get_dimnames}
\alias{test_FLQuant_get_dimnames2}
\alias{test_FLQuant_get_size}
\alias{test_FLQuant_get_nquant}
\alias{test_FLQuant_get_nyear}
\alias{test_FLQuant_get_nunit}
\alias{test_FLQuant_get_nseason}
\alias{test_FLQuant_get_narea}
\alias{test_FLQuant_get_niter}
\alias{test_FLQuant_get_first_age}
\alias{test_FLQuant_get_data_element}
\alias{test_FLQuant_get_const_single_index_accessor}
\alias{test_FLQuant_get_single_index_accessor}
\alias{test_FLQuant_set_single_index_accessor}
\alias{test_FLQuant_const_get_accessor}
\alias{test_FLQuant_get_accessor}
\alias{test_FLQuant_set_accessor}
\alias{test_FLQuant_get_const_indices_accessor}
\alias{test_FLQuant_get_indices_accessor}
\alias{test_FLQuant_set_indices_accessor}
\alias{test_FLQuant_get_all_iters}
\alias{test_FLQuant_set_data}
\alias{test_FLQuant_set_dimnames}
\alias{test_FLQuant_set_units}
\alias{test_FLQuant_subset}
\alias{test_FLQuant_neat_subset}
\alias{test_FLQuant_propagate_iters}
\alias{test_FLQuant_fill}
\alias{test_FLQuant_FLQuant_match_dims}
\alias{test_FLPar_to_FLQuant}
\alias{test_for_range}
\alias{test_for_range_const}
\alias{test_FLQuant_for_iterator}
\alias{test_FLQuant_for_iterator_const}
\alias{test_FLQuant_lambda}
\alias{test_DD_pow}
\alias{test_DAD_pow}
\alias{test_ADD_pow}
\alias{test_ADAD_pow}
\alias{test_input_subsetter_ADAD}
\alias{test_input_subsetter_DD}
\alias{test_input_subsetter_ADD}
\alias{test_input_subsetter_DAD}
\alias{test_year_season_to_timestep_FLQuant_double}
\alias{test_year_season_to_timestep_FLQuant_adouble}
\alias{test_year_season_to_timestep}
\alias{test_timestep_to_year_season_FLQuant_double}
\alias{test_timestep_to_year_season_FLQuant_adouble}
\alias{test_timestep_to_year_season}
\title{A metric fuck-tonne of them!}
\usage{
test_FLQuantAD_as_wrap(flq)

test_FLQuantAD_basic_constructor()

test_FLQuantAD_sexp_constructor(flq_sexp)

test_FLQuantAD_dim_constructor(nquant, nyear, nunit, nseason, narea, niter)

test_FLQuantAD_copy_constructor(flq)

test_FLQuantAD_copy_constructor2(flq1, quant, year, unit, season, area, iter,
  value)

test_FLQuantAD_assignment_operator(flq)

test_FLQuantAD_assignment_operator2(flq1, quant, year, unit, season, area, iter,
  value)

test_FLQuantAD_from_FLQuant_constructor(flq)

test_FLQuantAD_get_units(flq)

test_FLQuantAD_get_dim(flq)

test_FLQuantAD_get_dimnames(flq)

test_FLQuantAD_get_size(flq)

test_FLQuantAD_get_nquant(flq)

test_FLQuantAD_get_nyear(flq)

test_FLQuantAD_get_nunit(flq)

test_FLQuantAD_get_nseason(flq)

test_FLQuantAD_get_narea(flq)

test_FLQuantAD_get_niter(flq)

test_FLQuantAD_get_data_element(flq, quant, year, unit, season, area, iter)

test_FLQuantAD_get_const_single_index_accessor(flq, element)

test_FLQuantAD_get_single_index_accessor(flq, element)

test_FLQuantAD_set_single_index_accessor(flq, element, value)

test_FLQuantAD_const_get_accessor(flq, quant, year, unit, season, area, iter)

test_FLQuantAD_get_accessor(flq, quant, year, unit, season, area, iter)

test_FLQuantAD_set_accessor(flq, quant, year, unit, season, area, iter, value)

test_FLQuantAD_get_const_indices_accessor(flq, indices)

test_FLQuantAD_get_indices_accessor(flq, indices)

test_FLQuantAD_set_indices_accessor(flq, indices, value)

test_FLQuantAD_set_dimnames(flq, new_dimnames)

test_FLQuantAD_FLQuantAD_match_dims(flq1, flq2)

test_FLQuantAD_FLQuant_match_dims(flq1, flq2)

test_FLQuant_FLQuantAD_match_dims(flq1, flq2)

test_FLQuantAD_fill(flq, value)

test_FLQuantAD_fill_double(flq, value)

testFLCppAD(scalar)

FLPar_units(flp)

test_FLQuant_as_wrap(flq)

test_FLQuant_basic_constructor()

test_FLQuant_sexp_constructor(flq_sexp)

test_FLQuant_dim_constructor(nquant, nyear, nunit, nseason, narea, niter)

test_FLQuant_dim_value_constructor(nquant, nyear, nunit, nseason, narea, niter,
  value)

test_FLQuant_intvec_dim_constructor(dims)

test_FLQuant_intvec_dim_value_constructor(dims, value)

test_FLQuant_copy_constructor(flq)

test_FLQuant_copy_constructor2(flq1, quant, year, unit, season, area, iter,
  value)

test_FLQuant_assignment_operator(flq)

test_FLQuant_assignment_operator2(flq1, quant, year, unit, season, area, iter,
  value)

test_FLQuant_from_FLQuantAD_constructor(flqad)

test_FLQuant_get_data(flq)

test_FLQuant_get_units(flq)

test_FLQuant_get_dim(flq)

test_FLQuant_get_dimnames(flq)

test_FLQuant_get_dimnames2(flq)

test_FLQuant_get_size(flq)

test_FLQuant_get_nquant(flq)

test_FLQuant_get_nyear(flq)

test_FLQuant_get_nunit(flq)

test_FLQuant_get_nseason(flq)

test_FLQuant_get_narea(flq)

test_FLQuant_get_niter(flq)

test_FLQuant_get_first_age(flq)

test_FLQuant_get_data_element(flq, quant, year, unit, season, area, iter)

test_FLQuant_get_const_single_index_accessor(flq, element)

test_FLQuant_get_single_index_accessor(flq, element)

test_FLQuant_set_single_index_accessor(flq, element, value)

test_FLQuant_const_get_accessor(flq, quant, year, unit, season, area, iter)

test_FLQuant_get_accessor(flq, quant, year, unit, season, area, iter)

test_FLQuant_set_accessor(flq, quant, year, unit, season, area, iter, value)

test_FLQuant_get_const_indices_accessor(flq, indices)

test_FLQuant_get_indices_accessor(flq, indices)

test_FLQuant_set_indices_accessor(flq, indices, value)

test_FLQuant_get_all_iters(flq, quant, year, unit, season, area)

test_FLQuant_set_data(flq, data_in)

test_FLQuant_set_dimnames(flq, new_dimnames)

test_FLQuant_set_units(flq, new_units)

test_FLQuant_subset(flq, quant_min, quant_max, year_min, year_max, unit_min,
  unit_max, season_min, season_max, area_min, area_max, iter_min, iter_max)

test_FLQuant_neat_subset(flq, indices_min, indices_max)

test_FLQuant_propagate_iters(flq, iters)

test_FLQuant_fill(flq, value)

test_FLQuant_FLQuant_match_dims(flq1, flq2)

test_FLPar_to_FLQuant(flp)

test_for_range(flq, rn)

test_for_range_const(flq, rn)

test_FLQuant_for_iterator(flq, rn)

test_FLQuant_for_iterator_const(flq, rn)

test_FLQuant_lambda(flq1, flq2)

test_DD_pow(flq, power)

test_DAD_pow(flq, power)

test_ADD_pow(flq, power)

test_ADAD_pow(flq, power)

test_input_subsetter_ADAD(flq1, flq2, indices_min, indices_max)

test_input_subsetter_DD(flq1, flq2, indices_min, indices_max)

test_input_subsetter_ADD(flq1, flq2, indices_min, indices_max)

test_input_subsetter_DAD(flq1, flq2, indices_min, indices_max)

test_year_season_to_timestep_FLQuant_double(flq, year, season)

test_year_season_to_timestep_FLQuant_adouble(flqad, year, season)

test_year_season_to_timestep(flq, year, season)

test_timestep_to_year_season_FLQuant_double(flq, timestep)

test_timestep_to_year_season_FLQuant_adouble(flqad, timestep)

test_timestep_to_year_season(flq, timestep)
}
\arguments{
\item{flq}{something}

\item{flq_sexp}{something}

\item{nquant}{something}

\item{nyear}{something}

\item{nunit}{something}

\item{nseason}{something}

\item{narea}{something}

\item{niter}{something}

\item{flq1}{something}

\item{quant}{something}

\item{year}{something}

\item{unit}{something}

\item{season}{something}

\item{area}{something}

\item{iter}{something}

\item{value}{something}

\item{element}{something}

\item{indices}{something}

\item{new_dimnames}{something}

\item{flq2}{something}

\item{scalar}{Atmosblisters}

\item{flp}{something}

\item{dims}{something}

\item{flqad}{something}

\item{data_in}{something}

\item{new_units}{something}

\item{quant_min}{something}

\item{quant_max}{something}

\item{year_min}{something}

\item{year_max}{something}

\item{unit_min}{something}

\item{unit_max}{something}

\item{season_min}{something}

\item{season_max}{something}

\item{area_min}{something}

\item{area_max}{something}

\item{iter_min}{something}

\item{iter_max}{something}

\item{indices_min}{something}

\item{indices_max}{something}

\item{iters}{something}

\item{rn}{something}

\item{power}{do you have it?}

\item{timestep}{something}
}
\description{
A metric fuck-tonne of them!
}
