% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_FLQ_in_FLS}
\alias{test_FLQ_in_FLS}
\alias{test_FLStock_sexp_constructor}
\alias{test_FLStock_wrap}
\alias{test_FLStock_as}
\alias{test_FLStock_as_wrap}
\alias{test_FLStock_copy_constructor}
\alias{test_FLStock_assignment_operator}
\title{Killed by hippie powers}
\usage{
test_FLQ_in_FLS(flq)

test_FLStock_sexp_constructor(fls_sexp)

test_FLStock_wrap(fls_sexp)

test_FLStock_as(fls)

test_FLStock_as_wrap(fls)

test_FLStock_copy_constructor(fls1, quant, year, unit, season, area, iter,
  value)

test_FLStock_assignment_operator(fls1, quant, year, unit, season, area, iter,
  value)
}
\arguments{
\item{flq}{something}

\item{fls_sexp}{a parameter}

\item{fls}{something}

\item{fls1}{something}

\item{quant}{something}

\item{year}{something}

\item{unit}{something}

\item{season}{something}

\item{area}{something}

\item{iter}{something}

\item{value}{something}
}
\description{
Killed by hippie powers
}
