% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_fwdBiol_sexp_constructor}
\alias{test_fwdBiol_sexp_constructor}
\alias{test_fwdBiolAD_sexp_constructor}
\alias{test_fwdBiol_as_wrap}
\alias{test_fwdBiolAD_as_wrap}
\alias{test_fwdBiol_fwdSR_constructor}
\alias{test_fwdBiolAD_fwdSRAD_constructor}
\alias{test_fwdBiol_FLSR_bits_constructor}
\alias{test_fwdBiolAD_FLSR_residuals_constructor}
\alias{test_fwdBiolAD_FLSR_bits_constructor}
\alias{test_fwdBiol_copy_constructor}
\alias{test_fwdBiolAD_copy_constructor}
\alias{test_fwdBiol_copy_constructor2}
\alias{test_fwdBiolAD_copy_constructor2}
\alias{test_fwdBiol_assignment_operator}
\alias{test_fwdBiolAD_assignment_operator}
\alias{test_fwdBiol_assignment_operator2}
\alias{test_fwdBiolAD_assignment_operator2}
\alias{test_fwdBiol_const_get_accessors}
\alias{test_fwdBiol_const_get_accessors_subset}
\alias{test_fwdBiol_get_accessors}
\alias{test_fwdBiolAD_const_get_accessors}
\alias{test_fwdBiolAD_get_accessors}
\alias{test_fwdBiol_set_accessors}
\alias{test_fwdBiolAD_set_accessors}
\alias{fwdBiolAD_biomass_FLQ}
\alias{fwdBiolAD_biomass_subset}
\alias{test_fwdBiolAD_n_direct_set_accessor}
\alias{test_fwdBiolAD_n_direct_get_accessor}
\alias{test_fwdBiol_srp_timelag}
\alias{test_fwdBiols_as_wrap}
\alias{test_fwdBiolsAD_as_wrap}
\alias{test_fwdBiolsAD_fwdBiolAD_constructor}
\alias{test_fwdBiolsAD_const_get_single_index_accessor}
\alias{test_fwdBiolsAD_get_single_index_accessor}
\alias{test_fwdBiolsAD_const_get_value_accessor}
\alias{test_fwdBiolsAD_get_value_accessor}
\alias{test_fwdBiolsAD_set_single_index_accessor}
\alias{test_fwdBiolsAD_set_value_accessor}
\alias{test_fwdBiolsAD_copy_constructor}
\alias{test_fwdBiolsAD_assignment_operator}
\alias{test_fwdBiolsAD_const_iterator}
\alias{test_fwdBiolsAD_iterator}
\title{On black blessed wings}
\usage{
test_fwdBiol_sexp_constructor(flb_sexp)

test_fwdBiolAD_sexp_constructor(flb_sexp)

test_fwdBiol_as_wrap(fwdb)

test_fwdBiolAD_as_wrap(fwdb)

test_fwdBiol_fwdSR_constructor(flb_sexp, model_name, params, residuals,
  residuals_mult)

test_fwdBiolAD_fwdSRAD_constructor(flb_sexp, model_name, params, residuals,
  residuals_mult)

test_fwdBiol_FLSR_bits_constructor(flb_sexp, model_name, params, residuals,
  residuals_mult)

test_fwdBiolAD_FLSR_residuals_constructor(flb_sexp, residuals, residuals_mult)

test_fwdBiolAD_FLSR_bits_constructor(flb_sexp, model_name, params, residuals,
  residuals_mult)

test_fwdBiol_copy_constructor(fwdb1)

test_fwdBiolAD_copy_constructor(fwdb1)

test_fwdBiol_copy_constructor2(fwdb1, quant, year, unit, season, area, iter,
  value)

test_fwdBiolAD_copy_constructor2(fwdb1, quant, year, unit, season, area, iter,
  value)

test_fwdBiol_assignment_operator(fwdb1)

test_fwdBiolAD_assignment_operator(fwdb1)

test_fwdBiol_assignment_operator2(fwdb1, quant, year, unit, season, area, iter,
  value)

test_fwdBiolAD_assignment_operator2(fwdb1, quant, year, unit, season, area,
  iter, value)

test_fwdBiol_const_get_accessors(fwdb, quant, year, unit, season, area, iter)

test_fwdBiol_const_get_accessors_subset(fwdb, indices_min, indices_max)

test_fwdBiol_get_accessors(fwdb, quant, year, unit, season, area, iter)

test_fwdBiolAD_const_get_accessors(fwdb, quant, year, unit, season, area, iter)

test_fwdBiolAD_get_accessors(fwdb, quant, year, unit, season, area, iter)

test_fwdBiol_set_accessors(fwdb, quant, year, unit, season, area, iter, values)

test_fwdBiolAD_set_accessors(fwdb, quant, year, unit, season, area, iter,
  values)

fwdBiolAD_biomass_FLQ(fwdb)

fwdBiolAD_biomass_subset(fwdb, indices_min, indices_max)

test_fwdBiolAD_n_direct_set_accessor(fwdb, quant, year, unit, season, area,
  iter, value)

test_fwdBiolAD_n_direct_get_accessor(fwdb, quant, year, unit, season, area,
  iter)

test_fwdBiol_srp_timelag(fwb)

test_fwdBiols_as_wrap(biols_in)

test_fwdBiolsAD_as_wrap(biols_in)

test_fwdBiolsAD_fwdBiolAD_constructor(flb_sexp, residuals, residuals_mult)

test_fwdBiolsAD_const_get_single_index_accessor(fwbs, element)

test_fwdBiolsAD_get_single_index_accessor(fwbs, element)

test_fwdBiolsAD_const_get_value_accessor(fwbs, biol_no, quant, year, unit,
  season, area, iter)

test_fwdBiolsAD_get_value_accessor(fwbs, biol_no, quant, year, unit, season,
  area, iter)

test_fwdBiolsAD_set_single_index_accessor(fwbs, element, flb_sexp, residuals,
  residuals_mult)

test_fwdBiolsAD_set_value_accessor(fwbs, biol_no, quant, year, unit, season,
  area, iter, value)

test_fwdBiolsAD_copy_constructor(fwbs1, biol_no, dims, value)

test_fwdBiolsAD_assignment_operator(fwbs1, biol_no, dims, value)

test_fwdBiolsAD_const_iterator(fwbs)

test_fwdBiolsAD_iterator(fwbs, quant, year, unit, season, area, iter, value)
}
\arguments{
\item{flb_sexp}{a parameter}

\item{fwdb}{something}

\item{model_name}{something}

\item{params}{something}

\item{residuals}{something}

\item{residuals_mult}{something}

\item{fwdb1}{something}

\item{quant}{something}

\item{year}{something}

\item{unit}{something}

\item{season}{something}

\item{area}{something}

\item{iter}{something}

\item{value}{something}

\item{indices_min}{something}

\item{indices_max}{something}

\item{values}{something}

\item{fwb}{something}

\item{biols_in}{something}

\item{fwbs}{something}

\item{element}{something}

\item{biol_no}{something}

\item{fwbs1}{something}

\item{dims}{something}
}
\description{
On black blessed wings
}
