% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{[,fwdControl,ANY,ANY,ANY-method}
\alias{[,fwdControl,ANY,ANY,ANY-method}
\alias{[<-,fwdControl,ANY,ANY,vector-method}
\alias{[<-,fwdControl,ANY,ANY,ANY-method}
\alias{$,fwdControl-method}
\alias{$<-,fwdControl,vector-method}
\alias{$<-,fwdControl,AsIs-method}
\title{Set and replacement accessors for fwdControl}
\usage{
\S4method{[}{fwdControl,ANY,ANY,ANY}(x, i, j)

\S4method{[}{fwdControl,ANY,ANY,vector}(x, i, j, k, ...) <- value

\S4method{[}{fwdControl,ANY,ANY,ANY}(x, i, j, k, ...) <- value

\S4method{$}{fwdControl}(x, name)

\S4method{$}{fwdControl,vector}(x, name) <- value

\S4method{$}{fwdControl,AsIs}(x, name) <- value
}
\arguments{
\item{x}{A fwdControl object}

\item{i}{Row of both target and iters}

\item{j}{Third dimenions of iters}

\item{k}{The replacement.}

\item{...}{Some things.}

\item{value}{Replacement value}

\item{name}{Column name of target or value column of iters.}
}
\description{
We're Pastie to be Grill You
}
