% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_as_wrap_fwdControl}
\alias{test_as_wrap_fwdControl}
\alias{test_fwdControl_copy_constructor}
\alias{test_fwdControl_assignment_operator}
\alias{test_fwdControl_get_target}
\alias{test_fwdControl_get_ntarget}
\alias{test_fwdControl_get_niter}
\alias{test_fwdControl_get_nsim_target}
\alias{test_fwdControl_get_target_row}
\alias{test_fwdControl_get_target_rows}
\alias{test_fwdControl_get_target_int_col}
\alias{test_fwdControl_get_target_int_col2}
\alias{test_fwdControl_get_target_list_int_col}
\alias{test_fwdControl_get_target_list_int_col2}
\alias{test_fwdControl_get_target_num_col}
\alias{test_fwdControl_get_target_num_col2}
\alias{test_fwdControl_get_target_value}
\alias{test_fwdControl_get_target_value2}
\alias{test_fwdControl_get_target_quantity}
\alias{test_fwdControl_get_age_range}
\alias{test_fwdControl_get_FC}
\alias{test_fwdControl_get_B}
\alias{test_fwdControl_get_F}
\alias{test_fwdControl_get_FCB_nrow}
\alias{test_fwdControl_get_FCB_row_no}
\alias{test_fwdControl_get_FCB_nos}
\alias{test_fwdControl_shared_catch}
\title{Once around the sun}
\usage{
test_as_wrap_fwdControl(control)

test_fwdControl_copy_constructor(fc1)

test_fwdControl_assignment_operator(fc)

test_fwdControl_get_target(control)

test_fwdControl_get_ntarget(control)

test_fwdControl_get_niter(control)

test_fwdControl_get_nsim_target(control, target_no)

test_fwdControl_get_target_row(control, target_no, sim_target_no)

test_fwdControl_get_target_rows(control, target_no)

test_fwdControl_get_target_int_col(control, target_no, col)

test_fwdControl_get_target_int_col2(control, target_no, sim_target_no, col)

test_fwdControl_get_target_list_int_col(control, target_no, col)

test_fwdControl_get_target_list_int_col2(control, target_no, sim_target_no, col)

test_fwdControl_get_target_num_col(control, target_no, col)

test_fwdControl_get_target_num_col2(control, target_no, sim_target_no, col)

test_fwdControl_get_target_value(control, target_no, col)

test_fwdControl_get_target_value2(control, target_no, sim_target_no, col)

test_fwdControl_get_target_quantity(control, target_no, sim_target_no)

test_fwdControl_get_age_range(control, target_no, sim_target_no)

test_fwdControl_get_FC(control, biol_no)

test_fwdControl_get_B(control, fishery_no, catch_no)

test_fwdControl_get_F(control, biol_no)

test_fwdControl_get_FCB_nrow(control)

test_fwdControl_get_FCB_row_no(control, fishery_no, catch_no, biol_no)

test_fwdControl_get_FCB_nos(control, target_no, sim_target_no, relative, check)

test_fwdControl_shared_catch(control, biol_no)
}
\arguments{
\item{control}{something}

\item{fc1}{something}

\item{fc}{a parameter}

\item{target_no}{something}

\item{sim_target_no}{something}

\item{col}{something}

\item{biol_no}{something}

\item{fishery_no}{something}

\item{catch_no}{something}

\item{relative}{something}

\item{check}{something}
}
\description{
Once around the sun
}
