% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/constructors.R
\docType{class}
\name{fwdControl}
\alias{fwdControl}
\alias{fwdControl-class}
\alias{fwdControl-methods}
\title{A class for the targets and limits of a fishery projection.}
\usage{
\S4method{fwdControl}{data.frame,array}(target, iters, ...)
}
\description{
The desired targets, limits and time steps used in fishery projections can be 
specified by creating an object of class \code{fwdControl}.
}
\details{
[...]
}
\section{Slots}{

    \describe{
    \item{target}{The table of quantities and time steps used as target (\code{data.frame}).}
    \item{target}{The values and limits for each target quantity and time step (\code{array}).}
}
}

\section{Validity}{
 \describe{
    \item{VALIDITY}{Neque porro quisquam est qui dolorem ipsum.}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{


A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}
\examples{

# CREATE targets on fishing mortality ('f') by year

target <- data.frame(year=2000:2010, value=rlnorm(11), quantity='f')

fwc <- fwdControl(target=target)

# INSPECT fwdControl object

show(fwc)

# Construct from data.frame and array
fcn <- fwdControl(data.frame(year=2000:2005, quant='f', value=0.5))
}
\author{
Iago Mosqueira, Finlay Scott - EC JRC.
}
\seealso{
\link{data.frame}
}
\keyword{classes}

