% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_fwdSR_empty_constructor}
\alias{test_fwdSR_empty_constructor}
\alias{test_fwdSR_constructor_wrap}
\alias{test_fwdSR_get_nparams}
\alias{test_fwdSR_eval}
\alias{test_fwdSR_get_params}
\alias{test_fwdSR_copy_constructor}
\alias{test_fwdSR_assignment_operator}
\alias{test_fwdSR_predict_recruitment}
\alias{test_fwdSR_does_recruitment_happen}
\title{Leviathan}
\usage{
test_fwdSR_empty_constructor()

test_fwdSR_constructor_wrap(model_name, params, residuals, residuals_mult)

test_fwdSR_get_nparams(model_name, params, residuals, residuals_mult)

test_fwdSR_eval(model_name, params, residuals, residuals_mult, srp,
  param_indices)

test_fwdSR_get_params(model_name, params, residuals, residuals_mult,
  param_indices)

test_fwdSR_copy_constructor(model_name, params, residuals, residuals_mult, srp)

test_fwdSR_assignment_operator(model_name, params, residuals, residuals_mult,
  srp)

test_fwdSR_predict_recruitment(model_name, params, residuals, residuals_mult,
  srp, initial_params_indices)

test_fwdSR_does_recruitment_happen(model_name, params, residuals,
  residuals_mult, unit, year, season)
}
\arguments{
\item{model_name}{a parameter}

\item{params}{something}

\item{residuals}{something}

\item{residuals_mult}{something}

\item{srp}{something}

\item{param_indices}{something}

\item{initial_params_indices}{something}

\item{unit}{something}

\item{year}{something}

\item{season}{something}
}
\description{
Leviathan
}
