% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{guessfcb}
\alias{guessfcb}
\title{Generate an FCB matrix from FLBiols and FLFisheries}
\usage{
guessfcb(biols, fisheries)
}
\arguments{
\item{biols}{The FLBiols.}

\item{fisheries}{The FLFisheries.}
}
\description{
Tries to generate FCB matrix based on names.
Internal function. Ignore.
}
