% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_plugin.R
\name{inlineCxxPlugin}
\alias{inlineCxxPlugin}
\title{Make the plug in for inline Cxx}
\usage{
inlineCxxPlugin(...)
}
\arguments{
\item{...}{I have no idea.}
}
\description{
Make the plug in for inline Cxx
}
