% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/accessors.R
\docType{methods}
\name{iters<-}
\alias{iters<-}
\alias{iters,fwdControl-method}
\alias{iters<-,fwdControl,array-method}
\title{Access and replace the iters slot of the fwdControl}
\usage{
iters(object, ...) <- value

\S4method{iters}{fwdControl}(object)

\S4method{iters}{fwdControl,array}(object) <- value
}
\arguments{
\item{object}{The fwdControl.}

\item{...}{Other things.}

\item{value}{The iters array to replace the existing one with.}
}
\description{
Access and replace the iters slot of the fwdControl

Access the iters slot of the fwdControl

Set the iters slot of the fwdControl
}
