% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{make_skipjack_operatingModel}
\alias{make_skipjack_operatingModel}
\title{Create an operating model based on Indian Ocean skipjack tuna}
\usage{
make_skipjack_operatingModel(niters = 1000, sd = 0.1)
}
\arguments{
\item{niters}{The number of iterations.}

\item{sd}{The standard deviation when applying random lognormal noise to some of the slots.}
}
\value{
A list of objects for sending to C++
}
\description{
Two FLFishery objects with 1 FLCatch each fishing on a single FLBiolcpp.
The model has 4 seasons with 2 units.
Spawning happens in seasons 1 and 3.
}

