% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{make_test_operatingModel}
\alias{make_test_operatingModel}
\title{Make a test operating model from a single FLStock object}
\usage{
make_test_operatingModel(
  fls,
  FCB,
  nseasons = 1,
  recruitment_seasons = 1,
  recruitment_age = 1,
  niters = 1000,
  sd = 0.1
)
}
\arguments{
\item{fls}{The FLStock that the OM is based on}

\item{FCB}{The FCB matrix}

\item{nseasons}{The number of seasons}

\item{recruitment_seasons}{A vector of seasons in which recruitment occurs (1 - 4)}

\item{recruitment_age}{The age of recruitment to the fishery}

\item{niters}{The number of iterations.}

\item{sd}{The standard deviation when applying random lognormal noise to some of the slots.}
}
\value{
A list of objects for sending to C++
}
\description{
Number of fisheries, catches and biols are taken from the FCB argument.
}
