% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{make_test_operatingModel0}
\alias{make_test_operatingModel0}
\title{Create a complex annual test operating model}
\usage{
make_test_operatingModel0(niters = 1000)
}
\value{
A list of objects for sending to C++
}
\description{
Creates a test operating model for testing FLFishery / FLCatch / FLBiolcpp interactions.
Implements all possible type of FCB interactions.
Two FLFishery objects with 4 FLCatch objects between them, fishing 4 FLBiolcpp objects.
All objects are based on ple4.
}

