% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{make_test_operatingModel2}
\alias{make_test_operatingModel2}
\title{Create a simple annual test operating model}
\usage{
make_test_operatingModel2(niters = 1000)
}
\arguments{
\item{niters}{Number of iterations.}
}
\value{
A list of objects for sending to C++
}
\description{
Creates a test operating model for testing FLFishery / FLCatch / FLBiolcpp interactions.
Two FLFishery objects with 1 FLCatch objects each fishing on the same, single FLBiolcpp.
All objects are based on ple4.
}

