% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{make_test_operatingModel3}
\alias{make_test_operatingModel3}
\title{Create a very simple annual test operating model}
\usage{
make_test_operatingModel3(niters = 1000, sd = 0.1)
}
\arguments{
\item{niters}{Number of iterations.}

\item{sd}{Standard deviation for the noise.}
}
\value{
A list of objects for sending to C++
}
\description{
Creates the bits for a simple test operating model for projections.
A single FLFishery object with 1 FLCatch fishing on a single FLBiolcpp.
All objects are based on ple4.
}

