% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_posns_names}
\alias{match_posns_names}
\title{Change names of biols, catches and fisheries in the control object into integer positions}
\usage{
match_posns_names(trg, biol_names, fishery_catch_names)
}
\arguments{
\item{trg}{The target slot of a fwdControl object}

\item{biol_names}{A vector of names in the FLBiols objects}

\item{fishery_catch_names}{A named list - elements of list are vector of the catch names of each fishery}
}
\value{
The updated target slot
}
\description{
Change names of biols, catches and fisheries in the control object into integer positions
}
\details{
Before calling the C++ code it is necessary for the catch, fishery and biol columns (and their Rel) equivalents
to be integers. The user can specify by name. This function changes the name to integer position and throws an
error if the name does not match.
}
