% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biols}
\alias{biols}
\alias{flfs}
\title{An FLBiols object used in mixed fishery vignette}
\format{
An FLBiols object with two FLBiol objects.

An FLFisheries object with two FLFishery objects.
}
\usage{
biols

flfs
}
\description{
An FLBiols object with two FLBiol objects, one for plaice and one for sole.
These are used in the mixed fishery example vignette.

An FLFisheries object with two FLFishery objects, a beam trawl and a gill netter.
These are used in the mixed fishery example vignette.
}
\keyword{datasets}
