% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_operatingModel_empty_constructor}
\alias{test_operatingModel_empty_constructor}
\alias{test_operatingModel_full_constructor}
\alias{test_operatingModel_get_niter}
\alias{test_operatingModel_get_exp_z_pre_spwn}
\alias{test_operatingModel_f_prop_spwn_FLQ_subset}
\alias{test_operatingModel_SRP_FLQ_subset}
\alias{test_operatingModel_total_SRP_FLQ_subset}
\alias{test_operatingModel_calc_rec}
\alias{test_operatingModel_get_f_FCB}
\alias{test_operatingModel_get_f_FCB_subset}
\alias{test_operatingModel_get_f_B_subset}
\alias{test_operatingModel_get_f_B}
\alias{test_operatingModel_fbar_FCB}
\alias{test_operatingModel_fbar_B}
\alias{test_operatingModel_nunit_z_subset}
\alias{test_operatingModel_nunit_f_B_subset}
\alias{test_operatingModel_nunit_f_FCB_subset}
\alias{test_operatingModel_survivors}
\alias{test_operatingModel_project_biols}
\alias{test_operatingModel_project_fisheries}
\alias{test_operatingModel_run}
\alias{test_operatingModel_run2}
\alias{test_operatingModel_eval_om}
\alias{test_operatingModel_get_target_value_hat}
\alias{test_operatingModel_get_target_value_hat2}
\alias{test_operatingModel_get_target_value}
\alias{test_operatingModel_get_target_value2}
\alias{test_operatingModel_get_target_hat_indices}
\alias{test_operatingModel_catches_subset}
\alias{test_operatingModel_landings_subset}
\alias{test_operatingModel_discards_subset}
\alias{test_operatingModel_catch_n_subset}
\alias{test_operatingModel_landings_n_subset}
\alias{test_operatingModel_discards_n_subset}
\alias{test_operatingModel_ssb_start}
\alias{test_operatingModel_biomass_start}
\alias{test_operatingModel_ssb_end}
\alias{test_operatingModel_ssb_spawn}
\alias{test_operatingModel_ssb_flash}
\alias{test_operatingModel_biomass_flash}
\alias{test_operatingModel_biomass_end}
\alias{test_operatingModel_biomass_spawn}
\alias{test_operatingModel_spawn_before_fishing}
\alias{test_operatingModel_fishing_before_spawn}
\title{Blood mountain}
\usage{
test_operatingModel_empty_constructor()

test_operatingModel_full_constructor(flfs, biols, ctrl)

test_operatingModel_get_niter(flfs, biols, ctrl)

test_operatingModel_get_exp_z_pre_spwn(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_f_prop_spwn_FLQ_subset(flfs, biols, ctrl, fishery_no,
  biol_no, indices_min, indices_max)

test_operatingModel_SRP_FLQ_subset(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_total_SRP_FLQ_subset(flfs, biols, ctrl, biol_no,
  indices_min, indices_max)

test_operatingModel_calc_rec(flfs, biols, ctrl, biol_no, unit, rec_timestep)

test_operatingModel_get_f_FCB(flfs, biols, ctrl, fishery_no, catch_no, biol_no)

test_operatingModel_get_f_FCB_subset(flfs, biols, ctrl, fishery_no, catch_no,
  biol_no, indices_min, indices_max)

test_operatingModel_get_f_B_subset(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_get_f_B(flfs, biols, ctrl, biol_no)

test_operatingModel_fbar_FCB(flfs, biols, ctrl, fishery_no, catch_no, biol_no,
  indices_min, indices_max)

test_operatingModel_fbar_B(flfs, biols, ctrl, biol_no, indices_min, indices_max)

test_operatingModel_nunit_z_subset(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_nunit_f_B_subset(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_nunit_f_FCB_subset(flfs, biols, ctrl, fishery_no, catch_no,
  biol_no, indices_min, indices_max)

test_operatingModel_survivors(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_project_biols(flfs, biols, ctrl, timestep)

test_operatingModel_project_fisheries(flfs, biols, ctrl, timestep)

test_operatingModel_run(flfs, biols, ctrl, effort_mult_initial, indep_min,
  indep_max, nr_iters = 50L)

test_operatingModel_run2(flfs, biols, ctrl, effort_mult_initial, indep_min,
  indep_max, nr_iters = 50L)

test_operatingModel_eval_om(flfs, biols, ctrl, strquantity, fishery_no,
  catch_no, biol_no, indices_min, indices_max)

test_operatingModel_get_target_value_hat(flfs, biols, ctrl, target_no,
  sim_target_no)

test_operatingModel_get_target_value_hat2(flfs, biols, ctrl, target_no)

test_operatingModel_get_target_value(flfs, biols, ctrl, target_no,
  sim_target_no)

test_operatingModel_get_target_value2(flfs, biols, ctrl, target_no)

test_operatingModel_get_target_hat_indices(flfs, biols, ctrl, target_no,
  sim_target_no, relative)

test_operatingModel_catches_subset(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_landings_subset(flfs, flbs_list_sexp, ctrl, biol_no,
  indices_min, indices_max)

test_operatingModel_discards_subset(flfs, flbs_list_sexp, ctrl, biol_no,
  indices_min, indices_max)

test_operatingModel_catch_n_subset(flfs, flbs_list_sexp, ctrl, biol_no,
  indices_min, indices_max)

test_operatingModel_landings_n_subset(flfs, flbs_list_sexp, ctrl, biol_no,
  indices_min, indices_max)

test_operatingModel_discards_n_subset(flfs, flbs_list_sexp, ctrl, biol_no,
  indices_min, indices_max)

test_operatingModel_ssb_start(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_biomass_start(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_ssb_end(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_ssb_spawn(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_ssb_flash(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_biomass_flash(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_biomass_end(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_biomass_spawn(flfs, biols, ctrl, biol_no, indices_min,
  indices_max)

test_operatingModel_spawn_before_fishing(flfs, biols, ctrl, biol_no,
  indices_min, indices_max)

test_operatingModel_fishing_before_spawn(flfs, biols, ctrl, biol_no,
  indices_min, indices_max)
}
\arguments{
\item{flfs}{a parameter}

\item{biols}{something}

\item{ctrl}{something}

\item{biol_no}{something}

\item{indices_min}{something}

\item{indices_max}{something}

\item{fishery_no}{something}

\item{unit}{something}

\item{rec_timestep}{something}

\item{catch_no}{something}

\item{timestep}{something}

\item{effort_mult_initial}{something}

\item{indep_min}{something}

\item{indep_max}{something}

\item{nr_iters}{something}

\item{strquantity}{something}

\item{target_no}{something}

\item{sim_target_no}{something}

\item{relative}{something}

\item{flbs_list_sexp}{something}
}
\description{
Blood mountain
}
