% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{operatingModelRun}
\alias{operatingModelRun}
\title{Used to run the projections.
Internal use only.}
\usage{
operatingModelRun(flfs, biols, ctrl, effort_mult_initial, indep_min, indep_max,
  nr_iters = 50L)
}
\arguments{
\item{flfs}{FLFisheries.}

\item{biols}{List of the Biol bits.}

\item{ctrl}{fwdControl.}

\item{effort_mult_initial}{Initial effort multiplier.}

\item{indep_min}{Minimum independent solver value.}

\item{indep_max}{Maximum independent solver value.}

\item{nr_iters}{Maximum number of iterations for solver.}
}
\description{
Used to run the projections.
Internal use only.
}
