% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{parsefwdList}
\alias{parsefwdList}
\title{Parse the list argument to fwd to make a fwdControl object}
\usage{
parsefwdList(...)
}
\arguments{
\item{...}{Things}
}
\description{
Internal function that is not for internal consumption.
}
