% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,FLStock,fwdControl-method}
\title{plot method for FLStock, fwdControl}
\usage{
\S4method{plot}{FLStock,fwdControl}(x, y, fill = "#E69F00", ...)
}
\arguments{
\item{x}{FlStock object to plot}

\item{y}{fwdControl from which to extract year ranges}
}
\description{
plot method for FLStock, fwdControl
}
\section{Generic function}{
 plot(x, y)
}

\examples{
data(ple4)
control <- fwdControl(year=2000:2008, quant="f", value=0.3)
# No fwd projection took place, simply passing year range
plot(ple4, control)
}
