% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_user_function}
\alias{eval_user_function}
\alias{eval_user_gradient}
\title{Gahhhhhh}
\usage{
eval_user_function(xpsexp, params)

eval_user_gradient(xpsexp, params)
}
\arguments{
\item{xpsexp}{Function pointer}

\item{params}{params}
}
\description{
Gahhhhhh

Doesn't work
}
