% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_max3}
\alias{test_max3}
\alias{test_max2}
\alias{test_max}
\alias{test_standard_conditional}
\alias{test_cppad_conditional}
\title{Not needed}
\usage{
test_max3(indep)

test_max2(indep)

test_max(indep)

test_standard_conditional(indep)

test_cppad_conditional(indep)
}
\arguments{
\item{indep}{a parameter}
}
\description{
Not needed
}
