% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{propagate,fwdControl-method}
\alias{propagate,fwdControl-method}
\title{Propagate the fwdControl}
\usage{
\S4method{propagate}{fwdControl}(object, iter, fill.iter = TRUE)
}
\arguments{
\item{object}{A fwdControl object.}

\item{iter}{The number of iterations.}

\item{fill.iter}{Fill the new iters with original values (TRUE) or NA (FALSE)}
}
\description{
Change the nuber of iterations in the iter slot of the fwdControl.
}
