% Generated by roxygen2 (4.0.2): do not edit by hand
\name{random_FLBiol_generator}
\alias{random_FLBiol_generator}
\title{Generate randomly sized and filled FLBiol objects}
\usage{
random_FLBiol_generator(sd = 100, ...)
}
\arguments{
\item{sd}{The standard deviation of the random numbers. Passed to rnorm() Default is 100.}

\item{fixed_dims}{A vector of length 6 with the fixed length of each of the FLQuant dimensions. If any value is NA it is randomly set using the max_dims argument. Default value is rep(NA,6).}

\item{max_dims}{A vector of length 6 with maximum size of each of the FLQuant dimensions. Default value is c(5,5,5,4,4,10).}
}
\value{
An FLBiol
}
\description{
Generate an FLBiol of random size and filled with normally distributed random numbers with a mean of 0.
Used for automatic testing, particularly of the fwdBiol class in CPP.
}
\examples{
flb <- random_FLBiol_generator()
summary(flb)
}

