% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_FLBiolcpp_generator}
\alias{random_FLBiolcpp_generator}
\title{Generate randomly sized and filled FLBiolcpp objects}
\usage{
random_FLBiolcpp_generator(sd = 100, ...)
}
\arguments{
\item{sd}{The standard deviation of the random numbers. Passed to rnorm() Default is 100.}

\item{...}{Other arguments to pass to random_FLQuant_generator().}
}
\value{
An FLBiolcpp
}
\description{
Generate an FLBiolcpp of random size and filled with normally distributed random numbers with a mean of 0.
Used for automatic testing, particularly of the fwdBiol class in CPP.
}
\examples{
flb <- random_FLBiolcpp_generator()
summary(flb)
}
