% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_FLFisheries_generator}
\alias{random_FLFisheries_generator}
\title{Generate a randomly filled and sized FLFisheries object}
\usage{
random_FLFisheries_generator(min_fisheries = 2, max_fisheries = 5, ...)
}
\arguments{
\item{min_fisheries}{The minimum number of FLFisheries in the fisheries list. Default is 2.}

\item{max_fisheries}{The maximum number of FLFisheries in the fisheries list. Default is 5.}

\item{min_catches}{The minimum number of catches. Default is 2.}

\item{max_catches}{The maximum number of FLCatches in the catches list. Default is 5.}

\item{fixed_dims}{A vector of length 6 with the fixed length of each of the FLQuant dimensions. If any value is NA it is randomly set using the max_dims argument. Default value is rep(NA,6).}

\item{max_dims}{A vector of length 6 with maximum size of each of the FLQuant dimensions. Default value is c(5,5,5,4,4,10).}

\item{sd}{The standard deviation of the random numbers. Passed to rnorm() Default is 100.}
}
\value{
An FLFishery object
}
\description{
Generate a randomly sized FLFisheries object filled with normally distributed random numbers with a mean of 0.
Used for automatic testing, particularly of the FLFisheries_base<T> class in CPP.
}
\examples{
flf <- FLasher:::random_FLFishery_list_generator()
summary(flf)
flf <- random_FLFishery_generator(fixed_dims = c(NA,10,1,1,1,1))
lapply(flf, summary)
flf <- random_FLFishery_generator(fixed_dims = c(NA,10,1,1,1,1), max_dims = c(100,NA,NA,NA,NA,NA))
}

