% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_FLQuant_generator}
\alias{random_FLQuant_generator}
\title{Generate randomly sized and filled FLQuant objects}
\usage{
random_FLQuant_generator(fixed_dims = rep(NA, 6), min_dims = rep(1, 6),
  max_dims = pmax(min_dims, c(5, 10, 5, 4, 4, 5)), min_age_name = 1,
  sd = 100)
}
\arguments{
\item{fixed_dims}{A vector of length 6 with the fixed length of each of the FLQuant dimensions. If any value is NA it is randomly set using the max_dims argument. Default value is rep(NA,6).}

\item{min_dims}{A vector of length 6 with minimum size of each of the FLQuant dimensions. Default value is c(1,1,1,1,1,1).}

\item{max_dims}{A vector of length 6 with maximum size of each of the FLQuant dimensions. Default value is c(5,10,5,4,4,5).}

\item{min_age_name}{The name of the first age group.}

\item{sd}{The standard deviation of the random numbers. Passed to rnorm() Default is 100.}
}
\value{
An FLQuant
}
\description{
Generate a randomly or fixed sized FLQuant filled with normally distributed random numbers with a mean of 0.
Used for automatic testing.
All dimensions have at least length 2.
}
\examples{
flq <- random_FLQuant_generator()
dim(flq)
summary(flq)
flq <- random_FLQuant_generator(fixed_dims = c(NA,10,1,4,1,NA))
dim(flq)
summary(flq)
}

