% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_FLQuant_list_generator}
\alias{random_FLQuant_list_generator}
\title{Generate lists of randomly sized and filled FLQuant objects}
\usage{
random_FLQuant_list_generator(min_elements = 1, max_elements = 10, ...)
}
\arguments{
\item{min_elements}{The minimum number of elements in the list. Default is 1.}

\item{max_elements}{The maximum number of elements in the list. Default is 10.}

\item{...}{Other arguments to pass to random_FLQuant_generator(), e.g. those that fix the size of the objects.}
}
\value{
A list of FLQuant objects
}
\description{
Generate a list of FLQuant objects filled with normally distributed random numbers with a mean of 0.
FLQuant objects can be randomly sized, depening on arguments passed to random_FLQuant_generator().
Used for automatic testing, particularly of the FLQuant7_base<T> class in CPP.
}
\examples{
flq_list <- random_FLQuant_list_generator()
length(flq_list)
summary(flq_list)
lapply(flq_list, summary)
}
