% Generated by roxygen2 (4.0.2): do not edit by hand
\name{random_fwdBiols_list_generator}
\alias{random_fwdBiols_list_generator}
\title{Generates a list that can be passed to the CPP fwdBiols constructor}
\usage{
random_fwdBiols_list_generator(min_biols = 1, max_biols = 5, ...)
}
\arguments{
\item{min_biols}{The minimum number of fwdBiols in the list. Default is 1.}

\item{max_biols}{The maximum number of fwdBiols in the list. Default is 5.}

\item{fixed_dims}{A vector of length 6 with the fixed length of each of the FLQuant dimensions. If any value is NA it is randomly set using the max_dims argument. Default value is rep(NA,6).}

\item{max_dims}{A vector of length 6 with maximum size of each of the FLQuant dimensions. Default value is c(5,5,5,4,4,10).}

\item{sd}{The standard deviation of the random numbers. Passed to rnorm() Default is 100.}
}
\value{
A list object
}
\description{
The fwdBiols constructor takes a list (fwdbiols_list). Each element of fwdbiols_list is a list of:
FLBiol, SRR model name, SRR params, SRR timelag, SRR residuals and SRR residuals mult.
This function generates randomly filled FLBiol objects, of the same size.
Used for automatic testing, particularly of the fwdBiols<T> class in CPP.
}
\examples{
fwdBiols <- random_fwdBiols_list_generator()
}

