% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_fwdControl_generator}
\alias{random_fwdControl_generator}
\title{Random fwdControl object creator}
\usage{
random_fwdControl_generator(years = 1:round(runif(1, min = 2, max = 3)),
  nseasons = 2, max_nsim_target = 3, niters = round(runif(1, min = 5,
  max = 10)))
}
\arguments{
\item{years}{numeric vector of years in the control object. Default value is 1:random interger (max = 10).}

\item{nseasons}{number of seasons in the projection}

\item{max_nsim_target}{maximum number of simultaneous targets in each timestep}

\item{niters}{the number of iterations. Default number is random integer (max = 10).}
}
\value{
A fwdControl object
}
\description{
Creates a random fwdControl object for testing purposes
}
