% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simple_fisheries_project}
\alias{simple_fisheries_project}
\title{Simple projection with minimal checks}
\usage{
simple_fisheries_project(flfs, flb, flsr, f, f_spwn, sr_residuals,
  sr_residuals_mult, timesteps)
}
\arguments{
\item{flfs}{FLFisheries (with a single FLFishery with a single FLCatch)}

\item{flb}{FLBiol}

\item{f}{List of fishing mortality FLQuant objects (only a list of length 1 to start with)}

\item{f_spwn}{List of fishing timing FLQuant objects (only a list of length 1 to start with) - not used at the moment - part of the SSB calculation}

\item{sr_residuals}{FLQuant of residuals for the recruitment}

\item{sr_residuals_mult}{Are the residuals multiplicative (TRUE)  or additive (FALSE)?}

\item{timesteps}{Continuous sequence of integers (years and seasons)}
}
\value{
A list of FLFisheries and FLBiol objects
}
\description{
Given FLFisheries, FLBiol, FLSR, F and f.spwn, project over timesteps
No dimension checks are made!
Recruitment is annual only. Happens at start of the year. SSB is calculated in previous year (or years depending on recruitment age).
}

