% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skjBiol}
\alias{skjBiol}
\title{skjBiol}
\format{an \code{FLQuants} object for the biololgical slots and an \code{list} for the stock recruitment data}
\source{
IOTC
}
\usage{
skjBiol
}
\description{
Biological data for Indian Ocean skipjack tuna (4 seasons and 2 spawning units)
}
\author{
Iago Mosqueira
}

