% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_euclid_norm}
\alias{test_euclid_norm}
\alias{test_NR_quad_iters}
\alias{test_NR_linear}
\title{Metallic spheres}
\usage{
test_euclid_norm(xvec)

test_NR_quad_iters(coefs, initial_value, max_iters, indep_min, indep_max,
  tolerance)

test_NR_linear(initial_value, coefs, max_iters, indep_min, indep_max, tolerance)
}
\arguments{
\item{xvec}{A vector of doubles.}

\item{coefs}{a parameter}

\item{initial_value}{something}

\item{max_iters}{something}

\item{indep_min}{something}

\item{indep_max}{something}

\item{tolerance}{something}
}
\description{
Metallic spheres

Glamrock cops
}
