% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/window.R
\docType{methods}
\name{stf}
\alias{stf}
\alias{stf,FLStock-method}
\alias{stf,FLBiol-method}
\title{Prepare object for future projection}
\usage{
stf(object, ...)

\S4method{stf}{FLStock}(object, nyears = 3, wts.nyears = 3,
  fbar.nyears = wts.nyears, f.rescale = FALSE, arith.mean = TRUE,
  na.rm = TRUE, end = dims(object)$maxyear + nyears,
  disc.nyears = wts.nyears)

\S4method{stf}{FLBiol}(object, nyears = 3, wts.nyears = 3,
  arith.mean = TRUE, na.rm = TRUE, end = dims(object)$maxyear + nyears)
}
\arguments{
\item{object}{The object (FLStock or FLBiol)}

\item{...}{Other things.}

\item{nyears}{Number of years to extend the object}

\item{wts.nyears}{Number of years to average over to get the future mean weights at age.}

\item{fbar.nyears}{Number of years to average the F over (only used if object is an FLStock)}

\item{f.rescale}{Rescale F (TRUE or FALSE - default is FALSE)}

\item{arith.mean}{If TRUE the arithmetic mean is used. If FALSE the geometric mean is used. Default is TRUE.}

\item{na.rm}{For the mean function.}

\item{end}{My beautiful friend}
}
\description{
Similar to the old STF method in FLAssess.
Extends the object by a number of years and fill in the life history characteristics by taking a mean of the last few years.

stf for FLStock

stf for FLBiol
}
\examples{
data(ple4)

proj <- stf(ple4, 3)
}
