% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/accessors.R
\docType{methods}
\name{target}
\alias{target}
\alias{target<-}
\alias{target,fwdControl-method}
\alias{target<-,fwdControl,data.frame-method}
\title{Access and replace the target}
\usage{
target(object, ...)

target(object, ...) <- value

\S4method{target}{fwdControl}(object)

\S4method{target}{fwdControl,data.frame}(object) <- value
}
\arguments{
\item{object}{The fwdControl object}

\item{...}{Other things.}

\item{value}{The target object to replace the existing one with.}
}
\description{
Access and replace the target slot of a fwdControl object

Replace the target slot of a fwdControl object
}
