% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{targetOrder}
\alias{targetOrder}
\title{Get the order of targets in a fwdControl}
\usage{
targetOrder(target, iters)
}
\arguments{
\item{target}{The target.}

\item{iters}{The iters.}
}
\description{
Targets must be processed by FLasher in the correct order.
Internal function. Ignore.
}
