% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{test_FLFisheries_equal}
\alias{test_FLFisheries_equal}
\title{Tests if two FLFisheries objects are the same}
\usage{
test_FLFisheries_equal(flfs1, flfs2)
}
\description{
Tests each component seperately - allows flexibility
}

