% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{test_FLQuant_equal}
\alias{test_FLQuant_equal}
\title{Tests if two fwdQuant objects are the same}
\usage{
test_FLQuant_equal(flq1, flq2)
}
\description{
Tests each component seperately - allows flexibility
Just checks dims and values
}

