% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{test_fwdControl_equal}
\alias{test_fwdControl_equal}
\title{Tests if two fwdControl objects are the same}
\usage{
test_fwdControl_equal(fwc1, fwc2)
}
\description{
Tests each component seperately - allows flexibility
Avoids problems with FCB being an attribute
}

