% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcb_draw.R
\docType{methods}
\name{draw}
\alias{draw}
\alias{draw-method}
\alias{draw,basicBlock-method}
\alias{draw-basicBlock}
\alias{draw,biolBlock-method}
\alias{draw-biolBlock}
\alias{draw,fisheryBlock-method}
\alias{draw-fisheryBlock}
\alias{draw,catchBlock-method}
\alias{draw-catchBlock}
\alias{draw,FCBDrawing-method}
\alias{draw-FCBDrawing}
\alias{draw,matrix-method}
\alias{draw-matrix}
\alias{draw,fwdControl-method}
\alias{draw-fwdControl}
\title{Generic method for drawing components the FCB matrix}
\usage{
draw(object, ...)

\S4method{draw}{basicBlock}(object)

\S4method{draw}{biolBlock}(object)

\S4method{draw}{fisheryBlock}(object)

\S4method{draw}{catchBlock}(object)

\S4method{draw}{FCBDrawing}(object)

\S4method{draw}{matrix}(object, fisheryNames = NULL, catchNames = NULL,
  biolNames = NULL)

\S4method{draw}{fwdControl}(object, fisheryNames = NULL,
  catchNames = NULL, biolNames = NULL)
}
\arguments{
\item{object}{The object - link{basicBlock}, link{biolBlock}, link{fisheryBlock}, link{catchBlock}, link{matrix}, link{FCBDrawing}}

\item{...}{Other arguments}

\item{fisheryNames}{A vector of names of the FLFishery blocks}

\item{catchNames}{A vector of names of the FLCatch blocks}

\item{biolNames}{A vector of names of the FLBiol blocks}
}
\value{
Nothing. Just draws something
}
\description{
Used to draw individual components or the whole FCB matrix
}
\examples{
FCB <- matrix(c(1,1,1,1,2,2,2,1,2,2,2,3,2,2,4), nrow=5, byrow=TRUE)
draw(FCB)
}
