% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/accessors.R
\name{FCB}
\alias{FCB}
\alias{FCB,fwdControl-method}
\alias{FCB<-,fwdControl,matrix-method}
\title{Create and extract the FCB slot in fwdControl}
\usage{
FCB(object, ...) <- value

\S4method{FCB}{fwdControl}(object)

\S4method{FCB}{fwdControl,matrix}(object) <- value
}
\arguments{
\item{object}{Input object to construct or extract from.}

\item{...}{Extract input arguments}

\item{value}{Input matrix}
}
\description{
The FCB slot in fwdControl specifies the relationships between the catches
('C') taken by each fishery ('F') from different biological units ('B').
}
\details{
This slot is of class *matrix* and has thre columns, named 'F', 'C' and 'B',
and as many rows as relationships between the FLBiol(s) and FLFishery(ies)
objects the fwdControl refers to.
}
\examples{
control <- fwdControl()
# Access FCB slot
FCB(control)
# Assign to existing fwdControl
FCB(control) <- FCB(c(f=1, c=1, b=2), c(f=1, c=2, b=2))
}
