% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{fwdControl}
\alias{fwdControl}
\alias{fwdControl-class}
\alias{fwdControl-methods}
\title{A class for the targets and limits of a fishery projection.}
\description{
The desired targets, limits and time steps used in fishery projections can be
specified by creating an object of class \code{fwdControl}.
}
\details{
[...]
}
\section{Slots}{

    \describe{
    \item{target}{The table of values, quantities and time steps used as target (\code{fwdElement}).}
}
}

\section{Validity}{
 \describe{
    \item{VALIDITY}{Neque porro quisquam est qui dolorem ipsum.}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{


A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}
\examples{
# CREATE targets on fishing mortality ('f') by year

target <- data.frame(year=2000:2010, value=rlnorm(11), quantity='f')

fwc <- fwdControl(target=target)

# INSPECT fwdControl object

show(fwc)
}
\author{
Iago Mosqueira, Finlay Scott - EC JRC.
}
\seealso{
\link{fwdElement}
}
\keyword{classes}

