% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{get_FLQuant_elements}
\alias{get_FLQuant_elements}
\title{Return 1D vector of element indices of an FLQuant}
\usage{
get_FLQuant_elements(flq, indices_min, indices_max)
}
\arguments{
\item{flq}{The FLQuant}

\item{indices_min}{The min indices (integer vector, length 6)}

\item{indices_max}{The max indices (integer vector, length 6)}
}
\description{
Given an FLQuant and the indices range, returns the vector of indices
}

