% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{make_test_operatingModel2}
\alias{make_test_operatingModel2}
\title{Create a simple annual test operating model}
\usage{
make_test_operatingModel2(niters = 1000)
}
\value{
A list of objects for sending to C++
}
\description{
Creates a test operating model for testing FLFishery / FLCatch / FLBiol interactions.
Two FLFishery objects with 1 FLCatch objects each fishing on the same, single FLBiol.
All objects are based on ple4.
}

