% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,FLStock,fwdControl-method}
\alias{plot,FLStock,fwdControl-method}
\alias{plot,FLQuant,fwdControl-method}
\alias{plot,FLQuants,fwdControl-method}
\alias{plot,FLStocks,fwdControl-method}
\title{plot methods that highlight years in fwdControl}
\usage{
\S4method{plot}{FLStock,fwdControl}(x, y, fill = "#E69F00", ...)

\S4method{plot}{FLQuant,fwdControl}(x, y, fill = "#E69F00", ...)

\S4method{plot}{FLQuants,fwdControl}(x, y, fill = "#E69F00", ...)

\S4method{plot}{FLStocks,fwdControl}(x, y, fill = "#E69F00", ...)
}
\arguments{
\item{x}{FlStock object to plot}

\item{y}{fwdControl from which to extract year ranges}

\item{fill}{Colour to fill projection years background}

\item{...}{Any other argument to be passed to [ggplotFL::plot]}
}
\description{
plot methods that highlight years in fwdControl

plot method for FLStocks, fwdControl
}
\section{Generic function}{
 plot(x, y)
}

\examples{
data(ple4)
control <- fwdControl(year=2008:2017, quant="f", value=0.3)
# No fwd projection took place, simply passing year range
plot(ple4, control)
plot(ssb(ple4), control)
plot(FLQuants(SSB=ssb(ple4), F=fbar(ple4)), control)
data(ple4)
control <- fwdControl(year=2008:2017, quant="f", value=0.3)
# No fwd projection took place, simply passing year range
plot(FLStocks(PLE4=ple4), control)
}
