% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_FLCatches_generator}
\alias{random_FLCatches_generator}
\title{Generates an FLCatches object - a list of randomly sized and filled FLCatch objects}
\usage{
random_FLCatches_generator(min_catches = 2, max_catches = 5, ...)
}
\arguments{
\item{min_catches}{The minimum number of catches. Default is 2.}

\item{max_catches}{The maximum number of catches. Default is 5.}

\item{...}{Other arguments passed to random_FLQuant_generator().}
}
\value{
An FLCatches objects
}
\description{
Generates a list of identically sized FLCatch objects filled with normally distributed random numbers with a mean of 0.
Used for automatic testing, particularly of the FLCatches_base<T> class in CPP.
}
\examples{
flcs <- random_FLCatches_generator()
length(flcs)
summary(flcs)
lapply(flcs, summary)
}

