% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_FLFisheries_generator}
\alias{random_FLFisheries_generator}
\title{Generate a randomly filled and sized FLFisheries object}
\usage{
random_FLFisheries_generator(min_fisheries = 2, max_fisheries = 5, ...)
}
\arguments{
\item{min_fisheries}{The minimum number of FLFisheries in the fisheries list. Default is 2.}

\item{max_fisheries}{The maximum number of FLFisheries in the fisheries list. Default is 5.}

\item{...}{Other arguments to pass to random_FLFishery_generator().}
}
\value{
An FLFishery object
}
\description{
Generate a randomly sized FLFisheries object filled with normally distributed random numbers with a mean of 0.
Used for automatic testing, particularly of the FLFisheries_base<T> class in CPP.
}
\examples{
flf <- random_FLFisheries_generator(fixed_dims = c(NA,10,1,1,1,1))
}
