% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_FLFishery_generator}
\alias{random_FLFishery_generator}
\title{Generate a randomly filled and sized FLFishery object}
\usage{
random_FLFishery_generator(min_catches = 2, max_catches = 5, sd = 1,
  ...)
}
\arguments{
\item{min_catches}{The minimum number of catches. Default is 2.}

\item{max_catches}{The maximum number of FLCatches in the catches list. Default is 5.}

\item{sd}{Standard deviation of the randomly generated FLQuant slots.}

\item{...}{Other arguments passed to random_FLCatches_generator().}
}
\value{
An FLFishery object
}
\description{
Generate a randomly sized FLFishery object filled with normally distributed random numbers with a mean of 0.
Used for automatic testing, particularly of the FLFishery_base<T> class in CPP.
}
\examples{
flf <- random_FLFishery_generator(fixed_dims = c(NA,10,1,1,1,1))
lapply(flf, summary)
flf <- random_FLFishery_generator(fixed_dims = c(NA,10,1,1,1,1), max_dims = c(100,NA,NA,NA,NA,NA))
}
