% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_FLQuant_list_generator}
\alias{random_FLQuant_list_generator}
\title{Generate lists of randomly sized and filled FLQuant objects}
\usage{
random_FLQuant_list_generator(min_elements = 1, max_elements = 10, ...)
}
\arguments{
\item{min_elements}{The minimum number of elements in the list. Default is 1.}

\item{max_elements}{The maximum number of elements in the list. Default is 10.}

\item{fixed_dims}{A vector of length 6 with the fixed length of each of the FLQuant dimensions. If any value is NA it is randomly set using the max_dims argument. Default value is rep(NA,6).}

\item{max_dims}{A vector of length 6 with maximum size of each of the FLQuant dimensions. Default value is c(5,5,5,4,4,10).}

\item{sd}{The standard deviation of the random numbers. Passed to rnorm() Default is 100.}
}
\value{
A list of FLQuant objects
}
\description{
Generate a list of randomly sized FLQuant objects filled with normally distributed random numbers with a mean of 0.
Used for automatic testing, particularly of the FLQuant7_base<T> class in CPP.
}
\examples{
flq_list <- random_FLQuant_list_generator()
length(flq_list)
summary(flq_list)
lapply(flq_list, summary)
}

