% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{random_fwdBiols_list_generator}
\alias{random_fwdBiols_list_generator}
\title{Generates a list that can be passed to the CPP fwdBiols constructor}
\usage{
random_fwdBiols_list_generator(min_biols = 1, max_biols = 5, ...)
}
\arguments{
\item{min_biols}{The minimum number of fwdBiols in the list. Default is 1.}

\item{max_biols}{The maximum number of fwdBiols in the list. Default is 5.}

\item{...}{Other arguments passed to random_FLBiolcpp_generator().}
}
\value{
A list object
}
\description{
The fwdBiols constructor takes a list (fwdbiols_list). Each element of fwdbiols_list is a list of:
FLBiolcpp, SRR model name, SRR params, SRR timelag, SRR residuals and SRR residuals mult.
This function generates randomly filled FLBiolcpp objects. Objects may be of different sizes unless appropriate arguments to random_FLBiolcpp_generator() are specified.
Used for automatic testing, particularly of the fwdBiols<T> class in CPP.
}
\examples{
fwdBiols <- random_fwdBiols_list_generator()
}
