% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,fwdControl-method}
\alias{show,fwdControl-method}
\title{Show method for fwdControl}
\usage{
\S4method{show}{fwdControl}(object)
}
\arguments{
\item{object}{A fwdControl}
}
\description{
More Gills Less Fishcakes
}
