% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_NR_quad_iters}
\alias{test_NR_quad_iters}
\alias{test_NR_linear}
\title{Glamrock cops}
\usage{
test_NR_quad_iters(coefs, initial_value, max_iters, indep_min, indep_max,
  tolerance)

test_NR_linear(initial_value, coefs, max_iters, indep_min, indep_max, tolerance)
}
\arguments{
\item{coefs}{a parameter}

\item{initial_value}{something}

\item{max_iters}{something}

\item{indep_min}{something}

\item{indep_max}{something}

\item{tolerance}{something}
}
\description{
Glamrock cops
}
