% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{summary}
\alias{summary}
\title{summary method for fwdControl}
\usage{
\S4method{summary}{fwdControl}(object)
}
\description{
summary method for fwdControl
}
\examples{
control <- fwdControl(data.frame(year=rep(2010:2015, each=2),
  quant=c("f", "catch"), min=c(rbind(NA, 20000)), max=c(rbind(NA, 30000)),
  value=c(rbind(seq(1, 1.3, length=6), NA))))

summary(control)
}
